Reduce CPU limits/requests of deployments
With the previous default of 500m for cpu.requests and cpu.limits (where
m = milli), we were hitting the default CPU limit (2). I don’t think we
actually need that much CPU in the non-webservice deployments, so tune
those down a bit to keep the total below 2. (In fact, it’s exactly 1
right now, but that will presumably change once we introduce a
function-evaluator.)