Add function evaluator
This one doesn’t use the “tool” label, because we don’t want to mount
our entire home directory – only the subdirectory from the
function-evaluator. This way, malicious user-supplied JS code cannot
read the secrets of the wiki.

Trying to install dependencies of the package didn’t quite work –
mysteriously, npm doesn’t install the merge package, even though it’s
required by the kad package. (This happens locally as well, and I
haven’t figured out why.) As it happens, the latest master of kad
removed the merge dependency, so I instructed npm to install that
version of kad, and then told Git to ignore the changes to the package
files with `git update-index --assume-untracked`. This is the reason why
we don’t do `npm ci` in the script – I’ll just have to manually install
dependencies when the package.lock of function-evaluator changes.

Also, because we run this in a Node.js container, Python is not
available and attempts to evaluate Python code will simply fail. I think
that’s the best we can do for now (much later we could maybe run
multiple evaluators with different images? but the orchestrator doesn’t
support this at the moment as far as I’m aware).