Block external traffic to internal pods
None of these pods are supposed to see traffic from anywhere other than
within the tool-notwikilambda namespace, so add a network policy to
block ingress from elsewhere. Note that we need to limit it to a certain
label (I picked “networking: internal” but could also be something
else), so that it doesn’t also apply to the webservice pod (which needs
to be reachable externally).

Bug: T286724