Restart function orchestrator and evaluator without downtime
Previously, the function orchestrator and evaluator were unavailable
while they were updating themselves, because the ran in a simple loop
that was interrupted by the updates. We can do better: remove the loop
and have the orchestrator/evaluator run indefinitely after updating
themselves; add a readinessProbe to tell kubernetes when the update is
done and the pod can start serving traffic; and install a separate
cronjob that periodically rollout-restarts both deployments. Kubernetes
will then only route traffic to the new (restarted and updated) pods
once they’re ready, and keep the old pods alive until then.

This is largely based on a StackOverflow answer [1], but we don’t have
permission to create service accounts, roles or role bindings on
Toolforge, so we just mount the main Kubernetes credentials into the
container. A kubectl binary isn’t available in the container image, so I
downloaded one into ~/bin/ (though it turns out to be identical to the
one in /usr/bin/, too).

Note that the CronJob uses the version v1beta1; upstream, a version v1
is already available, but only since Kubernetes 1.21, and Toolforge is
still on 1.19.

[1]: https://stackoverflow.com/a/58378834/1420237