Change `npm ci` to `npm install`
`npm ci` is supposed to be faster than `npm install` because it does
less work, but I guess that only applies to a fresh install (because
`npm ci` starts out by deleting node_modules/) – with an existing
install, in a `webservice shell` (i.e. at 100% CPU), the wall-clock time
is 4m28s for `npm ci` and only 17s for `npm install`, and in the
deployemnt (with 200m cpu), the update to node12 failed to rollout at
all (the pod reached the time limit before becoming ready), whereas with
`npm install` it went without problems. (The timings are for the
function-evaluator, but I assume the function-orchestrator will be
pretty similar.)